Complete switcheroo
After a good night's sleep, I was finally able to fix the bugs that were
plaguing the game I kept in the back-up folder. Since I want the game to
be my submission for the assignment, I swapped it's file locations with
the files I had there before. Now you'll find the game in the main
directory and the other stuff in the backup folder. Welcome to the
arena!